Save Feature Points
Closes #10
The images of the map will always be the same, so there is no need to calculate the feature points of those images everytime.
Now the feature points of the images of the db maps are being stored in a pickle.
Feature Points use the KeyPoints class that can not be saved in a pickle. To save KeyPoints first convert them to a list. When loading them do not forget to convert them back. These functions are in the utils file.